# bound goroutine counts