--- a/processor/formatters_test.go
+++ b/processor/formatters_test.go
@@ -23,6 +23,6 @@ func BenchmarkFileSummerize(b *testing.B) {
 		close(fileSummaryJobQueue)
 		b.StartTimer()
 
-		fileSummerize(&fileSummaryJobQueue)
+		fileSummarize(fileSummaryJobQueue)
 	}
 }